+++ webviz_config/__init__.py
@@ -1,4 +1,9 @@
-from pkg_resources import get_distribution, DistributionNotFound
+try:
+    # Python 3.8+
+    from importlib.metadata import version, PackageNotFoundError  # type: ignore
+except ModuleNotFoundError:
+    # Python < 3.8
+    from importlib_metadata import version, PackageNotFoundError  # type: ignore
 
 from ._theme_class import WebvizConfigTheme
 from ._localhost_token import LocalhostToken
@@ -7,7 +12,7 @@ from ._plugin_abc import WebvizPluginABC
 from ._shared_settings_subscriptions import SHARED_SETTINGS_SUBSCRIPTIONS
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    __version__ = version("webviz-config")
+except PackageNotFoundError:
     # package is not installed
     pass