+++ webviz_config/themes/__init__.py
@@ -1,4 +1,9 @@
-import pkg_resources
+try:
+    # Python 3.8+
+    from importlib.metadata import entry_points  # type: ignore
+except ModuleNotFoundError:
+    # Python < 3.8
+    from importlib_metadata import entry_points  # type: ignore
 
 from .. import WebvizConfigTheme
 from ._default_theme import default_theme
@@ -9,7 +14,7 @@ installed_themes = {  # pylint: disable=invalid-name
 
 __all__ = ["installed_themes"]
 
-for entry_point in pkg_resources.iter_entry_points("webviz_config_themes"):
+for entry_point in entry_points().get("webviz_config_themes", []):
     theme = entry_point.load()
 
     globals()[entry_point.name] = theme